requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
huaweicloudsdkecs
huaweicloudsdkcore
